koimetrics/requirementx.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.1
bson==0.5.8
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.12
django-adminlte3==0.1.6
django-bootstrap4==1.1.1
djongo==1.3.1
idna==2.9
isodate==0.6.0
lxml==4.5.0
numpy==1.18.2
pandas==1.0.3
pkgconfig==1.5.1
psycopg2==2.8.5
pymongo==3.10.1
python-dateutil==2.8.1
python-decouple==3.3
python-tbk==0.2.2
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.2.4
urllib3==1.25.8
xmlsec==1.3.3
zeep==3.4.0
